server/requirements.txt
fastapi
uvicorn
axios
sqlalchemy
bcrypt
python-jose
passlib
gunicorn
sqlmodel
sqlalchemy
timescaledb
requests
python-decouple
aiosmtplib
psycopg[binary]
pydantic[email]
python-multipart
python-dotenv